pyproject.toml
[project]
name = "tessera"
version = "0.1.0"
description = "No-code multi-agent AI orchestration for full project generation. Like mosaic tiles, specialized agents coordinate to build complete software projects."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-litellm>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings[yaml]>=2.12.0",
    "rich>=13.0.0",
    "textual>=0.47.0",
    "typer>=0.9.0",
    "requests>=2.31.0",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "slack-sdk>=3.38.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pyyaml>=6.0.0",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
    "opentelemetry-exporter-otlp-proto-http>=1.38.0",
    "google-cloud-aiplatform>=1.38",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.0",
]

[project.scripts]
tessera = "tessera.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
cache-dir = ".cache/uv"

[tool.ruff]
line-length = 100
target-version = "py313"
cache-dir = ".cache/ruff"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.13"
strict = true
cache_dir = ".cache/mypy"

[tool.coverage.run]
data_file = ".cache/.coverage"

[tool.coverage.html]
directory = ".cache/coverage"

[tool.pytest.ini_options]
minversion = "9.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
cache_dir = ".cache/pytest"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src/tessera",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-branch",
    "--cov-fail-under=80",
]
markers = [
    "unit: Unit tests for individual components",
    "integration: Integration tests for component interactions",
    "slow: Tests that take significant time to run",
]
# Pytest 9.0 features
console_output_style = "progress-even-when-capture-no"
verbosity_assertions = 2
log_cli = false
log_cli_level = "INFO"
